+++ bluefin-internal/CHANGELOG.md
@@ -1,3 +1,8 @@
+* Restrict type parameter of `PrimStateEff` to `Effects`
+
+  This is technically a breaking change, but we did not enforce a
+  major version bump for it.
+
 # 0.4.2.0
 
 * Add `Bluefin.Internal.DslBuilderEff`